- Updated versions.

+++ src/main/scala/nl/wwbakker/services/dota/statistics/model/PlayerStats.scala
@@ -8,11 +8,11 @@ object PlayerStats {
   case class PlayerStatInfo(statGetter: PlayerStatGetter, friendlyName: String)
 
   private val statsMap = Map[String, PlayerStatInfo](
-    "kills" -> PlayerStatInfo(_.kills, "kills"),
-    "deaths" -> PlayerStatInfo(_.deaths, "deaths"),
-    "assists" -> PlayerStatInfo(_.assists, "assists"),
-    "lastHits" -> PlayerStatInfo(_.last_hits, "last hits"),
-    "denies" -> PlayerStatInfo(_.denies, "denies"),
+    "kills" -> PlayerStatInfo(_.kills.toFloat, "kills"),
+    "deaths" -> PlayerStatInfo(_.deaths.toFloat, "deaths"),
+    "assists" -> PlayerStatInfo(_.assists.toFloat, "assists"),
+    "lastHits" -> PlayerStatInfo(_.last_hits.toFloat, "last hits"),
+    "denies" -> PlayerStatInfo(_.denies.toFloat, "denies"),
     "gpm" -> PlayerStatInfo(p => p.total_gold.toFloat / (p.duration / 60f), "gold per minute"),
     "xpm" -> PlayerStatInfo(p => p.total_xp.toFloat / (p.duration / 60f), "experience per minute"),
     "duration" -> PlayerStatInfo(_.duration / 60f, "duration in minutes"),
